Create option to fully disable HTTPS certificate validation
Added the ability to completely disable HTTPS certificate validation using new settings and UI components. Introduced `TrustAllTrustManager` to facilitate this feature and updated the HTTP client configuration accordingly. This feature is necessary as most scanners use self-signed certs and this is usually okay as long as it is in the context of a private network. In the future, certificate pinning might be implemented